requirements.txt
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.3.3
google-api-core==2.14.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-language==2.11.1
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
langchain==0.0.339
numpy==1.26.2
pandas==2.1.3
streamlit==1.28.2
langchain
